+++ test/nodely/engine/applicative_test.clj
@@ -2,7 +2,11 @@
   (:require
    [clojure.core.async :as async]
    [clojure.test :refer :all]
+   [clojure.test.check.clojure-test :refer [defspec]]
+   [clojure.test.check.generators :as gen]
+   [clojure.test.check.properties :as prop]
    [criterium.core :as criterium]
+   [loom.alg :as alg]
    [matcher-combinators.matchers :as matchers]
    [matcher-combinators.test :refer [match?]]
    [nodely.data :as data]
@@ -11,7 +15,9 @@
    [nodely.engine.applicative.synchronous :as synchronous]
    [nodely.engine.core :as core]
    [nodely.engine.core-async.core :as nodely.async]
+   [nodely.engine.core-async.lazy-scheduling :as lazy-scheduling]
    [nodely.engine.schema :as schema]
+   [nodely.fixtures :as fixtures]
    [nodely.syntax :as syntax :refer [>leaf >value]]
    [nodely.syntax.schema :refer [yielding-schema]]
    [promesa.core :as p]
@@ -28,6 +34,15 @@
                                 :b ?b
                                 :c ?c})})
 
+(def test-env+delay-core-async {:a (>leaf (+ 1 2))
+                                :b (>leaf (do (Thread/sleep 1000)
+                                              (* ?a 2)))
+                                :c (>leaf (do (Thread/sleep 1000)
+                                              (* ?a 3)))
+                                :d (>leaf {:a ?a
+                                           :b ?b
+                                           :c ?c})})
+
 (def tricky-example {:x (data/value 1)
                      :y (data/value 2)
                      :a (data/value 3)
@@ -40,6 +55,23 @@
                                      (>leaf ?b)
                                      (>leaf ?c))
                      :z (>leaf ?w)})
+
+(def env-with-blocking-tag {:a (>leaf (Thread/currentThread))
+                            :b (syntax/blocking (>leaf (Thread/currentThread)))
+                            :a-name (>leaf (.getName ?a))
+                            :b-name (>leaf (.getName ?b))
+                            :c (>leaf (str ?a-name " " ?b-name))})
+
+(def env-with-nine-sleeps {:a (syntax/blocking (>leaf (do (Thread/sleep 1000) :a)))
+                           :b (syntax/blocking (>leaf (do (Thread/sleep 1000) :b)))
+                           :c (syntax/blocking (>leaf (do (Thread/sleep 1000) :c)))
+                           :d (syntax/blocking (>leaf (do (Thread/sleep 1000) :d)))
+                           :e (syntax/blocking (>leaf (do (Thread/sleep 1000) :e)))
+                           :f (syntax/blocking (>leaf (do (Thread/sleep 1000) :f)))
+                           :g (syntax/blocking (>leaf (do (Thread/sleep 1000) :g)))
+                           :h (syntax/blocking (>leaf (do (Thread/sleep 1000) :h)))
+                           :i (syntax/blocking (>leaf (do (Thread/sleep 1000) :i)))
+                           :z (>leaf (into #{} [?a ?b ?c ?d ?e ?f ?g ?h ?i]))})
 
 (def env-with-sequence {:a (>leaf [1 2 3])
                         :b (syntax/>sequence inc ?a)})
@@ -80,6 +112,19 @@
       (is (match? (matchers/within-delta 100000000 1000000000) time-ns))))
   (testing "tricky example"
     (is (match? 4 (applicative/eval-key tricky-example :z)))))
+
+(deftest eval-key-test-core-async
+  (testing "eval promise"
+    (is (match? 3 (applicative/eval-key test-env :c {::applicative/context core-async/context}))))
+  (testing "async works"
+    (let [[time-ns result] (criterium/time-body (applicative/eval-key test-env+delay-core-async
+                                                                      :d
+                                                                      {::applicative/context core-async/context}))]
+      (is (match? {:a 3 :b 6 :c 9} result))
+      (is (match? (matchers/within-delta 100000000 1000000000) time-ns))))
+  (testing "tricky example"
+    (is (match? 4 (applicative/eval-key tricky-example :z
+                                        {::applicative/context core-async/context})))))
 
 (deftest eval-test
   (testing "eval promise"
@@ -176,10 +221,46 @@
                           :value  3}
                          (ex-data
                           (s/with-fn-validation
-                            (applicative/eval-key env-with-failing-schema :c {::applicative/fvalidate schema/fvalidate
-                                                                              ::applicative/context core-async/context}))))))
+                            (applicative/eval-key env-with-failing-schema :c
+                                                  {::applicative/fvalidate schema/fvalidate
+                                                   ::applicative/context core-async/context}))))))
     (testing "async response is equal to sync response with async user channels"
       (is (= 7 (applicative/eval-key env+go-block :c {::applicative/context core-async/context}))))
     (testing "channel-leaf"
       (is (= 7 (applicative/eval-key env+channel-leaf :c {::applicative/context core-async/context}))))))
 
+(deftest core-async-blocking-eval-test
+  (testing "eval of a blocking tagged node will happen in the `async-thread-macro` worker pool"
+    (is (match? #"async-thread-macro-\d+"
+                (applicative/eval-key env-with-blocking-tag :b-name {::applicative/context core-async/context})))
+    (is (match? #"async-dispatch-\d+"
+                (applicative/eval-key env-with-blocking-tag :a-name {::applicative/context core-async/context})))))
+
+(deftest thread-sleeping-test-proves-thread-works-how-we-expect
+  (testing "actually only 8 threads in the async dispatch worker pool"
+    (is (match? 8
+                @@#'clojure.core.async.impl.exec.threadpool/pool-size)))
+  (testing "when we have 9 1-second blocking nodes in one environment, it can run in fewer than 2 seconds"
+    (testing "async version runs parallel when option is neglected"
+      (let [[nanosec-async _] (criterium/time-body (applicative/eval-key env-with-nine-sleeps :z {::applicative/context core-async/context}))]
+        (is (match? (matchers/within-delta 1000000000 1000000000)
+                    nanosec-async))))))
+
+(defspec does-not-blow-up-spec
+  (prop/for-all [env (fixtures/env-gen {})]
+                (applicative/eval-key env
+                                      (rand-nth (keys env))
+                                      {::applicative/context core-async/context})
+                true))
+
+(deftest compare-engines
+  (let [sample-env (gen/generate (fixtures/env-gen {:node-generator fixtures/scalar-gen
+                                                    :min-stages     20
+                                                    :max-stages     20}))
+        a-key      (last (alg/topsort (core/env->graph sample-env)))
+        [time-applicative res-applicative] (criterium/time-body (applicative/eval-key sample-env a-key {::applicative/context core-async/context}))
+        [time-lazy-sched res-lazy-sched]  (criterium/time-body (lazy-scheduling/eval-key sample-env a-key))]
+    (testing "results are the same"
+      (is (= res-applicative res-lazy-sched)))
+    (testing "applicative is faster than lazy scheduler"
+      (is (match? neg? (- time-applicative time-lazy-sched))))))